node-archives/flask-app/requirements.txt
Flask==2.3.3
pyaudio==0.2.13
vosk==0.3.29
transformers==4.36.2
torch==2.1.0
requests==2.31.0
python-dotenv==1.0.0
numpy<2.0  